if no valid auth method available, don't run test that require authentication

+++ dataone/inst/tests/test.MNode.R
@@ -57,12 +57,21 @@ test_that("MNode generateIdentifier()", {
     # Skip as this test requires authentication
     skip_on_cran()
     library(dataone)
-    cn <- CNode("SANDBOX2")
-    mn <- getMNode(cn, "urn:node:mnDemo2")
-    newid <- generateIdentifier(mn, "UUID")
-    cname <- class(newid)
-    expect_that(cname, matches("character"))
-    expect_that(newid, matches("urn:uuid:"))
+    cn <- CNode("STAGING")
+    mn <- getMNode(cn, "urn:node:mnStageUCSB2")  
+    # Suppress PKIplus, cert missing warnings
+    am <- AuthenticationManager()
+    warnLevel <- getOption("warn")
+    options(warn = -1)
+    authValid <- isAuthValid(am, mn)
+    options(warn = warnLevel)
+    if(authValid) {
+      if(getAuthMethod(am) == "cert" && grepl("apple-darwin", sessionInfo()$platform)) skip("Skip authenticatin w/cert on Mac OS X")
+      newid <- generateIdentifier(mn, "UUID")
+      cname <- class(newid)
+      expect_that(cname, matches("character"))
+      expect_that(newid, matches("urn:uuid:"))
+    }
 })
 test_that("MNode describe()", {
   library(dataone)
@@ -79,93 +88,94 @@ test_that("MNode create(), update(), archive(), and delete()", {
     library(datapackage)
     library(XML)
     #cn <- CNode("SANDBOX")
-    cn <- CNode("DEV2")
+    #cn <- CNode("DEV2")
+    cn <- CNode("STAGING")
     # mnDemo1 is api v1 on 20151208, but that could change
     # Use this v1 node to test with both a current token available
     # and a certificate.
     #mnId <- "urn:node:mnSandboxUCSB2"
-    mnId <- "urn:node:mnDevUCSB2"
+    #mnId <- "urn:node:mnDevUCSB2"
+    mnId <- "urn:node:mnStageUCSB2"
     mn <- getMNode(cn, mnId)
     am <- AuthenticationManager()
-    if (!isAuthValid(am, mn)) {
-      stop(sprintf("Valid DataONE authentication for the node %s with API version %s is required for this test.", mnId, mn@APIversion))
-    }
-    newid <- generateIdentifier(mn, "UUID")
-    cname <- class(newid)
-    expect_that(cname, matches("character"))
-    expect_that(newid, matches("urn:uuid:"))
-    
-    # Ensure the user is logged in before running the tests
-    # Set 'user' to authentication subject, if available, so we will have permission to change this object
-    user <- getAuthSubject(am)
-    # If subject isn't available from the current authentication method, then try
-    # the session configuration.
-    if (is.na(user)) {
-      # If session config doesn't have subject_dn set, then use the failback DN
-      user <- getOption("subject_dn")
-      expect_false(is.null(user) || is.na(user), info="This test requires that \"subject_dn\" is set, see \"dataone-overview\" vignette") 
-    }
-    expect_that(user, matches("cilogon|dataone"))
-    
-    # Create a data object, and convert it to csv format
-    testdf <- data.frame(x=1:10,y=11:20)
-    csvfile <- paste(tempfile(), ".csv", sep="")
-    write.csv(testdf, csvfile, row.names=FALSE)
-    
-    # Create SystemMetadata for the object
-    format <- "text/csv"
-    size <- file.info(csvfile)$size
-    sha1 <- digest(csvfile, algo="sha1", serialize=FALSE, file=TRUE)
-    # specify series id for this sysmeta. This will only be used if uploading to a DataONE v2 node
-    seriesId <- UUIDgenerate()
-    sysmeta <- new("SystemMetadata", identifier=newid, formatId=format, size=size, submitter=user, rightsHolder=user, checksum=sha1,
-                   originMemberNode=mn@identifier, authoritativeMemberNode=mn@identifier, seriesId=seriesId)
-    sysmeta <- addAccessRule(sysmeta, "public", "read")
-    expect_that(sysmeta@checksum, equals(sha1))
-    expect_that(sysmeta@submitter, equals(user))
-    expect_that(sysmeta@rightsHolder, equals(user))
-    expect_that(sysmeta@originMemberNode, equals(mn@identifier))
-    expect_that(sysmeta@authoritativeMemberNode, equals(mn@identifier))
-    
-    # Upload the data to the MN using create(), checking for success and a returned identifier
-    response <- create(mn, newid, csvfile, sysmeta)
-    expect_that(response, not(is_null()))
-    expect_that(xmlValue(xmlRoot(response)), matches(newid))
-    
-    # Update the object with a new version
-    updateid <- generateIdentifier(mn, "UUID")
-    testdf <- data.frame(x=1:20,y=11:30)
-    csvfile <- paste(tempfile(), ".csv", sep="")
-    write.csv(testdf, csvfile, row.names=FALSE)
-    size <- file.info(csvfile)$size
-    sha1 <- digest(csvfile, algo="sha1", serialize=FALSE, file=TRUE)
-    sysmeta@identifier <- updateid
-    sysmeta@size <- size
-    sysmeta@checksum <- sha1
-    sysmeta@obsoletes <- newid
-    response <- update(mn, newid, csvfile, updateid, sysmeta)
-    expect_that(xmlValue(xmlRoot(response)), matches(updateid))
-    updsysmeta <- getSystemMetadata(mn, updateid)
-    expect_that(class(updsysmeta)[1], matches("SystemMetadata"))
-    expect_that(updsysmeta@obsoletes, matches(newid))
-    
-    # Now get the sysmeta using the seriesId, if supported
-    if(mn@APIversion >= "v2") {
-      headSysmeta <- getSystemMetadata(mn, seriesId)
-      expect_that(class(headSysmeta)[1], matches("SystemMetadata"))
-      expect_that(updsysmeta@identifier, matches(headSysmeta@identifier))
-    }
+    # Suppress PKIplus, cert missing warnings
+    warnLevel <- getOption("warn")
+    options(warn = -1)
+    authValid <- isAuthValid(am, mn)
+    options(warn = warnLevel)
+    if (authValid) {
+      if(getAuthMethod(am) == "cert" && grepl("apple-darwin", sessionInfo()$platform)) skip("Skip authenticatin w/cert on Mac OS X")
+      newid <- generateIdentifier(mn, "UUID")
+      cname <- class(newid)
+      expect_that(cname, matches("character"))
+      expect_that(newid, matches("urn:uuid:"))
       
-    # Archive the object
-    response <- archive(mn, newid)
-    expect_that(response, matches(newid))
-    newsysmeta <- getSystemMetadata(mn, newid)
-    expect_that(class(newsysmeta)[1], matches("SystemMetadata"))
-    expect_that(newsysmeta@archived, is_true())
-    
-    #TODO: delete the object
-    #response <- delete(mn, newid)
-    
+      # Set 'user' to authentication subject, if available, so we will have permission to change this object
+      user <- getAuthSubject(am)
+      # If subject isn't available from the current authentication method, then try
+      # R options
+      if (is.na(user) || user == "public") {
+        user <- getOption("subject_dn")
+        if(is.null(user) || is.na(user)) skip("This test requires that you set options(subject_dn = \"<your identity>\")")
+      }
+      
+      expect_that(user, matches("cilogon|dataone"))
+      # Create a data object, and convert it to csv format
+      testdf <- data.frame(x=1:10,y=11:20)
+      csvfile <- paste(tempfile(), ".csv", sep="")
+      write.csv(testdf, csvfile, row.names=FALSE)
+      
+      # Create SystemMetadata for the object
+      format <- "text/csv"
+      size <- file.info(csvfile)$size
+      sha1 <- digest(csvfile, algo="sha1", serialize=FALSE, file=TRUE)
+      # specify series id for this sysmeta. This will only be used if uploading to a DataONE v2 node
+      seriesId <- UUIDgenerate()
+      sysmeta <- new("SystemMetadata", identifier=newid, formatId=format, size=size, submitter=user, rightsHolder=user, checksum=sha1,
+                     originMemberNode=mn@identifier, authoritativeMemberNode=mn@identifier, seriesId=seriesId)
+      sysmeta <- addAccessRule(sysmeta, "public", "read")
+      expect_that(sysmeta@checksum, equals(sha1))
+      expect_that(sysmeta@submitter, equals(user))
+      expect_that(sysmeta@rightsHolder, equals(user))
+      expect_that(sysmeta@originMemberNode, equals(mn@identifier))
+      expect_that(sysmeta@authoritativeMemberNode, equals(mn@identifier))
+      
+      # Upload the data to the MN using create(), checking for success and a returned identifier
+      response <- create(mn, newid, csvfile, sysmeta)
+      expect_that(response, not(is_null()))
+      expect_that(xmlValue(xmlRoot(response)), matches(newid))
+      
+      # Update the object with a new version
+      updateid <- generateIdentifier(mn, "UUID")
+      testdf <- data.frame(x=1:20,y=11:30)
+      csvfile <- paste(tempfile(), ".csv", sep="")
+      write.csv(testdf, csvfile, row.names=FALSE)
+      size <- file.info(csvfile)$size
+      sha1 <- digest(csvfile, algo="sha1", serialize=FALSE, file=TRUE)
+      sysmeta@identifier <- updateid
+      sysmeta@size <- size
+      sysmeta@checksum <- sha1
+      sysmeta@obsoletes <- newid
+      response <- update(mn, newid, csvfile, updateid, sysmeta)
+      expect_that(xmlValue(xmlRoot(response)), matches(updateid))
+      updsysmeta <- getSystemMetadata(mn, updateid)
+      expect_that(class(updsysmeta)[1], matches("SystemMetadata"))
+      expect_that(updsysmeta@obsoletes, matches(newid))
+      
+      # Now get the sysmeta using the seriesId, if supported
+      if(mn@APIversion >= "v2") {
+        headSysmeta <- getSystemMetadata(mn, seriesId)
+        expect_that(class(headSysmeta)[1], matches("SystemMetadata"))
+        expect_that(updsysmeta@identifier, matches(headSysmeta@identifier))
+      }
+      
+      # Archive the object
+      response <- archive(mn, newid)
+      expect_that(response, matches(newid))
+      newsysmeta <- getSystemMetadata(mn, newid)
+      expect_that(class(newsysmeta)[1], matches("SystemMetadata"))
+      expect_that(newsysmeta@archived, is_true())
+    }
 })
 
 test_that("MNode create() works for large files", {
@@ -177,8 +187,8 @@ test_that("MNode create() works for large files", {
     library(dataone)
     library(digest)
     library(httr)
-    cn <- CNode("SANDBOX2")
-    mn <- getMNode(cn, "urn:node:mnDemo2")
+    cn <- CNode("STAGING")
+    mn <- getMNode(cn, "urn:node:mnStageUCSB2")
     newid <- generateIdentifier(mn, "UUID")
     cname <- class(newid)
     expect_that(cname, matches("character"))
@@ -186,46 +196,52 @@ test_that("MNode create() works for large files", {
     # Ensure the user is logged in before running the tests
     # Set 'user' to authentication subject, if available, so we will have permission to change this object
     am <- AuthenticationManager()
-    if (!isAuthValid(am, cn)) {
-      stop(sprintf("Valid DataONE authentication is required for this test."))
+    # Suppress PKIplus, cert missing warnings
+    warnLevel <- getOption("warn")
+    options(warn = -1)
+    authValid <- isAuthValid(am, mn)
+    options(warn = warnLevel)
+    if (authValid) {
+      if(getAuthMethod(am) == "cert" && grepl("apple-darwin", sessionInfo()$platform)) skip("Skip authenticatin w/cert on Mac OS X")
+      # Set 'subject' to authentication subject, if available, so we will have permission to change this object
+      subject <- getAuthSubject(am)
+      # If subject isn't available from the current authentication method, then try
+      # the R options.
+      if (is.na(subject) || user == "public") {
+        subject <- getOption("subject_dn")
+        if(is.null(subject) || is.na(subject)) skip("This test requires that you set options(subject_dn = \"<your identity>\")")
+      }
+      
+      expect_that(subject, matches("cilogon|dataone"))
+      # TODO: Create a large data object using fallocate through a system call (only on linux)
+      csvfile <- 'testdata.csv'
+      csvsize <- '4G'
+      system(paste("fallocate -l", csvsize, csvfile))
+      
+      # On a mac, use truncate instead (slower, and produces a sparse file)
+      #csvsize <- '4294967296'    # truncate needs the size arg in bytes
+      #system(paste("truncate", csvfile, csvsize))
+      
+      # Create SystemMetadata for the object
+      format <- "text/csv"
+      size <- file.info(csvfile)$size
+      sha1 <- digest(csvfile, algo="sha1", serialize=FALSE, file=TRUE)
+      sysmeta <- new("SystemMetadata", identifier=newid, formatId=format, size=size, submitter=user, rightsHolder=user, checksum=sha1, originMemberNode=mn@identifier, authoritativeMemberNode=mn@identifier)
+      sysmeta <- addAccessRule(sysmeta, "public", "read")
+      expect_that(sysmeta@checksum, equals(sha1))
+      expect_that(sysmeta@submitter, equals(user))
+      expect_that(sysmeta@rightsHolder, equals(user))
+      expect_that(sysmeta@originMemberNode, equals(mn@identifier))
+      expect_that(sysmeta@authoritativeMemberNode, equals(mn@identifier))
+      
+      # Upload the data to the MN using create(), checking for success and a returned identifier
+      response <- create(mn, newid, csvfile, sysmeta)
+      expect_that(response, not(is_null()))
+      expect_that(xmlValue(xmlRoot(response)), matches(newid)) 
+      
+      # Remove the big data file we created locally
+      unlink(csvfile)
     }
-    subject <- getAuthSubject(am)
-    # If subject isn't available from the current authentication method, then try
-    # the session configuration.
-    if (is.na(subject)) {
-      subject <- getOption("subject_dn")
-      expect_false(is.null(user) || is.na(user), info="This test requires that \"subject_dn\" is set, see \"dataone-overview\" vignette") 
-    }    
-    
-    expect_that(user, matches("cilogon|dataone"))
-    # TODO: Create a large data object using fallocate through a system call (only on linux)
-    csvfile <- 'testdata.csv'
-    csvsize <- '4G'
-    system(paste("fallocate -l", csvsize, csvfile))
-    
-    # On a mac, use truncate instead (slower, and produces a sparse file)
-    #csvsize <- '4294967296'    # truncate needs the size arg in bytes
-    #system(paste("truncate", csvfile, csvsize))
-    
-    # Create SystemMetadata for the object
-    format <- "text/csv"
-    size <- file.info(csvfile)$size
-    sha1 <- digest(csvfile, algo="sha1", serialize=FALSE, file=TRUE)
-    sysmeta <- new("SystemMetadata", identifier=newid, formatId=format, size=size, submitter=user, rightsHolder=user, checksum=sha1, originMemberNode=mn@identifier, authoritativeMemberNode=mn@identifier)
-    sysmeta <- addAccessRule(sysmeta, "public", "read")
-    expect_that(sysmeta@checksum, equals(sha1))
-    expect_that(sysmeta@submitter, equals(user))
-    expect_that(sysmeta@rightsHolder, equals(user))
-    expect_that(sysmeta@originMemberNode, equals(mn@identifier))
-    expect_that(sysmeta@authoritativeMemberNode, equals(mn@identifier))
-    
-    # Upload the data to the MN using create(), checking for success and a returned identifier
-    response <- create(mn, newid, csvfile, sysmeta)
-    expect_that(response, not(is_null()))
-    expect_that(xmlValue(xmlRoot(response)), matches(newid)) 
-    
-    # Remove the big data file we created locally
-    unlink(csvfile)
 })
 
 test_that("MNode getPackage() works", {
@@ -234,6 +250,16 @@ test_that("MNode getPackage() works", {
   cn <- CNode("SANDBOX2")
   mn <- getMNode(cn, "urn:node:mnDemo2")
   resMapPid <- "urn:uuid:62febde3-5e7b-47b8-97a9-a874ffc9a180"
+  #resMapPid <- "resourceMap_hpackage-test.1.1"
+  am <- AuthenticationManager()
+  warnLevel <- getOption("warn")
+  options(warn = -1)
+  authValid <- isAuthValid(am, mn)
+  options(warn = warnLevel)
+  if (authValid) {
+    if(getAuthMethod(am) == "cert" && grepl("apple-darwin", sessionInfo()$platform)) skip("Skip authenticatin w/cert on Mac OS X")
+  }
+  
   bagitFile <- getPackage(mn, id=resMapPid)
   expect_true(!is.null(bagitFile))
   expect_true(file.exists(bagitFile))
@@ -245,53 +271,44 @@ test_that("MNode getPackage() works", {
   
 })
 
-test_that("Obsoleting an object with updateSystemMetadata() works",{
+test_that("updateSystemMetadata() works",{
   skip_on_cran()
   library(dataone)
   # Create a csv file for the data object
   testdf <- data.frame(x=1:10,y=11:20)
   csvfile <- tempfile(pattern = "file", tmpdir = tempdir(), fileext = ".csv")
   write.csv(testdf, csvfile, row.names=FALSE)
-  mnId <- "urn:node:mnSandboxUCSB2"
-  d1c <- D1Client(env="SANDBOX", mNodeid=mnId)
-  # Set 'user' to authentication subject, if available, so we will have permission to change this object
+  mnId <- "urn:node:mnStageUCSB2"
+  d1c <- D1Client(env="STAGING", mNodeid=mnId)
+  
   am <- AuthenticationManager()
-  if (!isAuthValid(am, d1c@mn)) {
-    warning(sprintf("Valid DataONE authentication is required for this test."))
-  }
-  subject <- getAuthSubject(am)
-  # If subject isn't available from the current authentication method, then try
-  # the session configuration.
-  if (is.na(subject)) {
-    subject <- getOption("subject_dn")
-    # If option "subject_dn" not set, exit 
-    if (is.null(subject)) {
-      warning(sprintf("This test requires setting the \"subject_dn\" option. Please see \"dataone-overview\" vignette."))
-      skip("Need to set subject_dn to run this test.")
+  warnLevel <- getOption("warn")
+  options(warn = -1)
+  authValid <- isAuthValid(am, d1c@mn)
+  options(warn = warnLevel)
+  if (authValid) {
+    if(getAuthMethod(am) == "cert" && grepl("apple-darwin", sessionInfo()$platform)) skip("Skip authenticatin w/cert on Mac OS X")
+    # Set 'user' to authentication subject, if available, so we will have permission to change this object
+    subject <- getAuthSubject(am)
+    # If subject isn't available from the current authentication method, then try
+    # R options
+    if (is.na(subject) || subject == "public") {
+      subject <- getOption("subject_dn")
+      if(is.null(subject) || is.na(subject)) skip("This test requires that you set options(subject_dn = \"<your identity>\")")
     }
+    
+    do1 <- new("DataObject", format="text/csv", user=subject, mnNodeId=mnId, filename=csvfile)
+    # Set replication off, to prevent the bug of serialNumber increasing due to replication bug
+    uploadDataObject(d1c, do1, replicate=FALSE, public=TRUE)
+    id1 <- getIdentifier(do1)
+    md1 <- getSystemMetadata(d1c@mn, id1)
+    expect_false(is.null(md1))
+    id <- "uid=slaughter,o=NCEAS,dc=ecoinformatics,dc=org"
+    md1 <- addAccessRule(md1, id, "read")
+    expect_true(updateSystemMetadata(d1c@mn, md1@identifier, md1))
+    md1New  <- getSystemMetadata(d1c@mn, id1)
+    # Did the access policy get updated?
+    ap <- md1New@accessPolicy
+    expect_true(ap[ap$subject==id,"subject"] == id)
   }
-  
-  do1 <- new("DataObject", format="text/csv", user=subject, mnNodeId=mnId, filename=csvfile)
-  # Set replication off, to prevent the bug of serialNumber increasing due to replication bug
-  uploadDataObject(d1c, do1, replicate=FALSE, public=TRUE)
-  do2 <- new("DataObject", format="text/csv", user=subject, mnNodeId=mnId, filename=csvfile)
-  # Set replication off, to prevent the bug of serialNumber increasing due to replication bug
-  uploadDataObject(d1c, do2, replicate=FALSE, public=TRUE)
-  id1 <- getIdentifier(do1)
-  id2 <- getIdentifier(do2)
-  md1 <- getSystemMetadata(d1c@mn, id1)
-  expect_false(is.null(md1))
-  md2 <- getSystemMetadata(d1c@mn, id2)
-  expect_false(is.null(md2))
-  md1@obsoletedBy <- id2
-  md2@obsoletes <- id1
-  
-  expect_true(updateSystemMetadata(d1c@mn, md1@identifier, md1))
-  expect_true(updateSystemMetadata(d1c@mn, md2@identifier, md2))
-  
-  tstmd1 <- getSystemMetadata(d1c@mn, id1)
-  expect_equal(tstmd1@obsoletedBy, id2)
-  tstmd2 <- getSystemMetadata(d1c@mn, id2)
-  expect_equal(tstmd2@obsoletes, id1)
-  
 })
